Use a paramlist to represent a struct ucred
The ucred structure differs on Linux, FreeBSD, OpenBSD and NetBSD.
I haven't checked Mac OS X or Solaris yet but they are probably different
too.

To accomodate these differences, credrecv/1 returns the raw binary. cred/1
is a convenience function to parse the binary into a parameter list based
on the platform. For example, Linux is the only OS that returns the PID and
all of the OS'es return the egid of the process execpt FreeBSD.

Still only tested on Linux/32-bit/little-endian/arm.