cred/1: convert a proplist to a struct ucred
Create an OS specific struct ucred from a list for use with explicit
cred passing.

The default uid/gid are set to 0 since there doesn't seem to be a way
to retrieve the current user (without use of an NIF). This means that
if the ucred structure is created with the default parameters and the
user is not root, credsend/1 will fail.