Portability fixes for FreeBSD
fd passing now works on FreeBSD.

Chnage macros for constants to functions. The values differ between
OS'es. SOL_SOCKET is a constant in procket.hrl. It will need to be
changed to return a value based on OS.

Fix the padding for structs on FreeBSD. The ucred structure on Linux is
called cmsgcred on FreeBSD. Further work is required to support
credential passing.