Support encode/decode of (c)msghdr
Have inverse functions for creating the cmsghdr and msghdr.

The functions are not correct yet:

* fdsend/recv is a tuple, credsend/recv is an atom

Should support an optional argument, since fdsend/recv default to 1
file descriptor.

* mismatched types in encode/decode

fdsend takes a list of integers, an integer or a binary encoding the
integers.

fdrecv takes an integer respresenting the number of fd's to return.

msg({fdrecv, N}) returns a list of integers.

msg(credrecv) returns a proplist.

* before receiving a credential, the caller must manually call
setsockopt/2 on the socket