Update the README
The gen_unix interface is still changing. The API is still not correct.
It also requires the caller to poll the socket, to manually make calls
to e.g., setsockopt/2 and to track state.

The plan is to rename gen_unix -> unixsock. unixsock will have the
socket primitives and will be a simple function interface.

gen_unix will be a gen_server. gen_unix will track the fd's, poll them
(either using inert or by sleeping and spinning on the socket) and close
the socket/delete the socket file on termination.